Parse tax table data into hash
This change adds the ability to parse the tax tables into a `Hash` with
the key being the first column - zipcode and the value a decimal
representation of the second column - the tax rate.

This change also moves the fixture files into separate folders, which ensures
we can properly test tables w/ header rows and tables without header rows.